Add LUA Script and Fix Bugs
Converted the Python script to LUA so that users don't have to have a Python installation to use this script. Fixed defaults being the wrong values. Fixed a bug where changing the source while the script was disabled would cause it to be unable to be disabled and would just halve the speed when disabled.